+++ src/EngineCore/Entities/EntityInstaller.cs
@@ -24,6 +24,7 @@
 using Castle.MicroKernel.Registration;
 using Castle.MicroKernel.SubSystems.Configuration;
 using Castle.Windsor;
+using Sovereign.EngineUtil.IoC;
 
 namespace Sovereign.EngineCore.Entities
 {
@@ -31,14 +32,18 @@ namespace Sovereign.EngineCore.Entities
     /// <summary>
     /// IoC installer for the entity infrastructure.
     /// </summary>
-    public class EntityInstaller : IWindsorInstaller
+    public sealed class EntityInstaller : IWindsorInstaller
     {
 
         public void Install(IWindsorContainer container, IConfigurationStore store)
         {
-            /* EntityManager. */
             container.Register(Component.For<EntityManager>()
                 .LifestyleSingleton());
+
+            container.Register(EngineClasses.EngineAssemblies()
+                .BasedOn<IEntityFactory>()
+                .WithServiceDefaultInterfaces()
+                .LifestyleTransient());
         }
 
     }